% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acctran.R
\name{acctran}
\alias{acctran}
\alias{parsimony_edgelength}
\title{Assign edge length to tree}
\usage{
acctran(tree, data)

parsimony_edgelength(tree, data)
}
\arguments{
\item{tree}{a tree, i.e. an object of class pml}

\item{data}{an object of class phyDat}
}
\value{
a tree with edge length.
}
\description{
\code{parsimony_edgelength} and \code{acctran} assign edge length to a tree where
the edge length is the number of mutations. \code{parsimony_edgelengths}
assigns edge lengths using a joint reconstruction based on the sankoff
algorithm. Ties are broken at random and trees can be multifurating.
\code{acctran} is based on the fitch algorithm and is faster. However trees
need to be bifurcating and ties are split.
}
\examples{
# lower number of iterations for the example, to run time less than 5 sec.
treeRatchet <- pratchet(Laurasiatherian, minit=5, k=5, trace=0)
# assign edge length (number of substitutions)
treeRatchet <- parsimony_edgelength(treeRatchet, Laurasiatherian)
plot(midpoint(treeRatchet))
add.scale.bar(0,0, length=100)
}
\seealso{
\code{\link{pratchet}}
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
