% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFreq.R
\name{baseFreq}
\alias{baseFreq}
\alias{glance.phyDat}
\alias{composition_test}
\alias{summary.phyDat}
\alias{print.summary.phyDat}
\title{Summaries of alignments}
\usage{
baseFreq(obj, freq = FALSE, all = FALSE, drop.unused.levels = FALSE)

\method{glance}{phyDat}(x, ...)

composition_test(obj)

\method{summary}{phyDat}(object, ...)

\method{print}{summary.phyDat}(x, ..., digits = max(3L, getOption("digits") -
  3L))
}
\arguments{
\item{freq}{logical, if 'TRUE', frequencies or counts are returned otherwise
proportions}

\item{all}{all a logical; if all = TRUE, all counts of bases, ambiguous
codes, missing data, and alignment gaps are returned as defined in the
contrast.}

\item{drop.unused.levels}{logical, drop unused levels}

\item{...}{further arguments passed to or from other methods.}

\item{object, obj, x}{as object of class phyDat}

\item{digits}{minimal number of significant digits.}
}
\value{
\code{baseFreq} returns a named vector and \code{glance} a one row
\code{data.frame}.
}
\description{
\code{baseFreq} computes the frequencies (absolute or relative) of the states
from a sample of sequences.
\code{glance} computes some useful information about the alignment.
\code{composition\_test} computes a \eqn{\chi^2}-test testing if the state
composition for a species differs.
}
\examples{

data(Laurasiatherian)
data(chloroplast)
# base frequencies
baseFreq(Laurasiatherian)
baseFreq(Laurasiatherian, all=TRUE)
baseFreq(Laurasiatherian, freq=TRUE)
baseFreq(chloroplast)
glance(Laurasiatherian)
glance(chloroplast)
composition_test(Laurasiatherian)[1:10,]
}
\seealso{
\code{\link{phyDat}, \link[ape]{base.freq}, \link{glance}}
}
\author{
Klaus Schliep
}
