% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap.pml}
\alias{bootstrap.pml}
\alias{bootstrap.phyDat}
\title{Bootstrap}
\usage{
bootstrap.pml(x, bs = 100, trees = TRUE, multicore = FALSE,
  mc.cores = NULL, tip.dates = NULL, ...)

bootstrap.phyDat(x, FUN, bs = 100, multicore = FALSE, mc.cores = NULL,
  jumble = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{pml} or \code{phyDat}.}

\item{bs}{number of bootstrap samples.}

\item{trees}{return trees only (default) or whole \code{pml} objects.}

\item{multicore}{logical, whether models should estimated in parallel.}

\item{mc.cores}{The number of cores to use during bootstrap. Only supported
on UNIX-alike systems.}

\item{tip.dates}{A named vector of sampling times associated to the
tips/sequences. Leave empty if not estimating tip dated phylogenies.}

\item{\dots}{further parameters used by \code{optim.pml} or
\code{plot.phylo}.}

\item{FUN}{the function to estimate the trees.}

\item{jumble}{logical, jumble the order of the sequences.}
}
\value{
\code{bootstrap.pml} returns an object of class \code{multi.phylo}
or a list where each element is an object of class \code{pml}. \code{plotBS}
returns silently a tree, i.e. an object of class \code{phylo} with the
bootstrap values as node labels. The argument \code{BStrees} is optional and
if not supplied the tree with labels supplied in the \code{node.label} slot.
}
\description{
\code{bootstrap.pml} performs (non-parametric) bootstrap analysis and
\code{bootstrap.phyDat} produces a list of bootstrapped data sets.
}
\details{
It is possible that the bootstrap is performed in parallel using the future
package, see example below.
}
\examples{

\dontrun{
data(Laurasiatherian)
dm <- dist.hamming(Laurasiatherian)
tree <- NJ(dm)
# NJ
set.seed(123)
NJtrees <- bootstrap.phyDat(Laurasiatherian,
     FUN=function(x)NJ(dist.hamming(x)), bs=100)
treeNJ <- plotBS(tree, NJtrees, "phylogram")

# Maximum likelihood
fit <- pml(tree, Laurasiatherian)
fit <- optim.pml(fit, rearrangement="NNI")
set.seed(123)
# compute in parallel
future::plan(multisession, workers = 2)
bs <- bootstrap.pml(fit, bs=100, optNni=TRUE)
future::plan(sequential)

treeBS <- plotBS(fit$tree,bs)

# Maximum parsimony
treeMP <- pratchet(Laurasiatherian)
treeMP <- acctran(treeMP, Laurasiatherian)
set.seed(123)
BStrees <- bootstrap.phyDat(Laurasiatherian, pratchet, bs = 100)
treeMP <- plotBS(treeMP, BStrees, "phylogram")
add.scale.bar()

# export tree with bootstrap values as node labels
# write.tree(treeBS)
}

}
\references{
\bibshow{*, Felsenstein1985, Penny1985, Penny1986, Lemoine2018}
}
\seealso{
\code{\link{optim.pml}}, \code{\link{pml}},
\code{\link[future]{plan}},
\code{\link[ape]{plot.phylo}}, \code{\link{maxCladeCred}}
\code{\link[ape]{nodelabels}},\code{\link{consensusNet}} and
\code{\link{SOWH.test}} for parametric bootstrap
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
\keyword{cluster}
