\name{chloroplast}
\alias{chloroplast}
\docType{data}
\title{ Chloroplast alignment }
\description{
Amino acid alignment of 15 genes of 19 different chloroplast.}
\usage{data(yeast)}

%\references{
%Nisbet R.E.R, Schliep K., Steel M.A., Knapp M. Howe C.J. and Lockhart P.J. (2009)
%The slop of rocks and clocks
%\emph{Potato growers weekly}, \bold{425}(6960): 798--804
%}    

\examples{
data(chloroplast)
chloroplast
}
\keyword{datasets}
