% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusNet.R
\name{consensusNet}
\alias{consensusNet}
\title{Computes a consensusNetwork from a list of trees}
\usage{
consensusNet(obj, prob = 0.3, ...)
}
\arguments{
\item{obj}{An object of class multiPhylo.}

\item{prob}{the proportion a split has to be present in all trees to be
represented in the network.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
\code{consensusNet} returns an object of class networx.  This is
just an intermediate to plot phylogenetic networks with igraph.
}
\description{
\code{consensusNet} computes a split network, i.e. an object of class
\code{networx} from a list of trees, i.e. an class of class
\code{multiPhylo} \bibcitep{Holland2004}.
}
\examples{

data(Laurasiatherian)
set.seed(1)
bs <- bootstrap.phyDat(Laurasiatherian, FUN = function(x)nj(dist.hamming(x)),
    bs=50)
cnet <- consensusNet(bs, 0.3)
plot(cnet, angle=-60)
\dontrun{
library(rgl)
open3d()
plot(cnet, type = "3D", show.tip.label=FALSE, show.nodes=TRUE)
plot(cnet, type = "equal angle", show.edge.label=TRUE)

tmpfile <- normalizePath(system.file(
              "extdata/trees/RAxML_bootstrap.woodmouse", package="phangorn"))
trees <- read.tree(tmpfile)
cnet_woodmouse <- consensusNet(trees, 0.3)
plot(cnet_woodmouse, type = "equal angle", show.edge.label=TRUE)
}

}
\references{
\bibshow{*}
}
\seealso{
\code{\link{splitsNetwork}}, \code{\link{neighborNet}},
\code{\link{lento}}, \code{\link{distanceHadamard}},
\code{\link{plot.networx}}, \code{\link{maxCladeCred}}
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
\keyword{hplot}
