% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.score.R
\name{delta.score}
\alias{delta.score}
\title{Computes the \eqn{\delta} score}
\usage{
delta.score(x, arg = "mean", ...)
}
\arguments{
\item{x}{an object of class \code{phyDat}}

\item{arg}{Specifies the return value, one of "all", "mean" or "sd"}

\item{...}{further arguments passed through \code{dist.hamming}}
}
\value{
A vector containing the \eqn{\delta} scores.
}
\description{
Computes the treelikeness
}
\examples{

data(yeast)
hist(delta.score(yeast, "all"))

}
\references{
\bibshow{*, Holland2002, Gray2010}
}
\seealso{
\code{\link{dist.hamming}}
}
\author{
Alastair Potts and Klaus Schliep
}
\keyword{cluster}
