% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadamard.R
\name{ldfactorial}
\alias{ldfactorial}
\alias{dfactorial}
\title{Arithmetic Operators}
\usage{
ldfactorial(x)

dfactorial(x)
}
\arguments{
\item{x}{a numeric scalar or vector}
}
\value{
\code{dfactorial(x)} returns the double factorial, that is
\eqn{x\!\! = 1 * 3 * 5 * \ldots * x } and \code{ldfactorial(x)} is the
natural logarithm of it.
}
\description{
double factorial function
}
\examples{

dfactorial(1:10)

}
\seealso{
\code{\link[base:Special]{factorial}},
\code{\link[ape]{howmanytrees}}
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
\keyword{classif}
