% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.gamma.R
\name{discrete.gamma}
\alias{discrete.gamma}
\alias{discrete.beta}
\alias{plot_gamma_plus_inv}
\alias{plotRates}
\title{Discrete Gamma and Beta distribution}
\usage{
discrete.gamma(alpha, k, w = NULL)

discrete.beta(shape1, shape2, k)

plot_gamma_plus_inv(w = NULL, g = NULL, shape = 1, inv = 0, k = 4,
  discrete = TRUE, cdf = TRUE, append = FALSE, xlab = "x",
  ylab = ifelse(cdf, "F(x)", "f(x)"), xlim = NULL, verticals = FALSE,
  edge.length = NULL, site.rate = "gamma", ...)

plotRates(obj, cdf.color = "blue", main = "", rug = FALSE, xlim = NULL,
  append = FALSE, ...)
}
\arguments{
\item{alpha}{Shape parameter of the gamma distribution.}

\item{k}{Number of intervals of the discrete gamma distribution.}

\item{w}{proportion of rate g.}

\item{shape1, shape2}{non-negative parameters of the Beta distribution.}

\item{g}{rates of discrete distribution.}

\item{shape}{Shape parameter of the gamma distribution.}

\item{inv}{Proportion of invariable sites.}

\item{discrete}{logical whether to plot discrete (default) or continuous pdf
or cdf.}

\item{cdf}{logical whether to plot the cumulative distribution function
or density / probability function.}

\item{append}{logical; if TRUE only add to an existing plot.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{xlim}{the x limits of the plot.}

\item{verticals}{logical; if TRUE, draw vertical lines at steps.}

\item{edge.length}{Total edge length (sum of all edges in a tree).}

\item{site.rate}{Indicates what type of gamma distribution to use. Options
are "gamma" (Yang 1994), "gamma_quadrature" using Laguerre quadrature
approach of Felsenstein (2001) or  "variable_weight_gamma".}

\item{\dots}{Further arguments passed to or from other methods.}

\item{obj}{an object of class pml}

\item{cdf.color}{color of the cdf.}

\item{main}{a main title for the plot.}

\item{rug}{logical; if TRUE a \code{\link[graphics]{rug}} is added to the
plot.}
}
\value{
\code{discrete.gamma} returns a matrix.
}
\description{
\code{discrete.gamma} internally used for the likelihood computations in
\code{pml} or \code{optim.pml}. The functions \code{plot_gamma_plus_inv}
and \code{plotRates} are useful in EDA and understand how discrete rate
models work and where it limitations are \bibcitep{Ferretti2026}.
\code{plotRates} plots the cumulative distribution function based on the
parsimony score for the data and the discrete rate model.
}
\examples{
discrete.gamma(1, 4)

old.par <- par(no.readonly = TRUE)
par(mfrow = c(2,1))
plot_gamma_plus_inv(shape=2, discrete = FALSE, cdf=FALSE)
plot_gamma_plus_inv(shape=2, append = TRUE, cdf=FALSE)

plot_gamma_plus_inv(shape=2, discrete = FALSE)
plot_gamma_plus_inv(shape=2, append = TRUE)
par(old.par)

data(Laurasiatherian)
fit <- pml_bb(Laurasiatherian, "JC+G(4)", rearrangement = "none")
plotRates(fit)
}
\references{
\bibshow{*}
}
\seealso{
\code{\link{pml.fit}, \link{stepfun}, \link{pgamma}, \link{pbeta}}
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
\keyword{distribution}
\keyword{hplot}
