% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_lengths.R
\name{draw_ci}
\alias{draw_ci}
\title{Draw Confidences Intervals on Phylogenies}
\usage{
draw_ci(CI, col95 = "#FF00004D", col50 = "#0000FF4D", height = 0.7,
  legend = TRUE, ...)
}
\arguments{
\item{CI}{output from [chronosCI()] or a similar matrix.}

\item{col95}{colour used for the 95\% intervals; by default: transparent
red.}

\item{col50}{colour used for the 50\% intervals; by default: transparent
blue.}

\item{height}{the height of the boxes.}

\item{legend}{a logical value.}

\item{\dots}{arguments passed to [graphics::legend()]}
}
\description{
This is a low-level plotting command to draw the confidence
intervals on the node of a tree as rectangles with coloured backgrounds.
}
\details{
The matrix \code{CI} must have four rows and as many columns as the
number of nodes of the tree. The first and fourth rows give the lower and
upper bounds of the 95% confidence intervals. The second and third rows
give the lower and upper bounds of the 50% confidence intervals. The Trees
should to be rooted, either ultrametric or tip dated.
}
\examples{
\dontrun{
data("Laurasiatherian")
dm <- dist.hamming(Laurasiatherian)
tree <- upgma(dm)
set.seed(123)
trees <- bootstrap.phyDat(Laurasiatherian,
                          FUN=function(x)upgma(dist.hamming(x)), bs=100)
                          tree <- plotBS(tree, trees, "phylogram")
X <- edge_lengths_matrix(tree, trees, TRUE)
Y <- apply(X, 2, \(x, probs)quantile(na.omit(x), probs),
           probs=c(.025,.25,.75,.975))
tree <- plotBS(tree, trees, "phylogram")
draw_ci(Y[,-c(1:47)], height = .8)
}
}
\seealso{
[plot.phylo, plotBS]
}
\author{
Emmanuel Paradis, Santiago Claramunt, Joseph Brown, Klaus Schliep
}
\keyword{aplot}
