% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_as_state.R
\name{gap_as_state}
\alias{gap_as_state}
\alias{gap_as_ambiguous}
\title{Treat gaps as a state}
\usage{
gap_as_state(obj, gap = "-", ambiguous = "?")

gap_as_ambiguous(obj, gap = "-")
}
\arguments{
\item{obj}{An object of class phyDat.}

\item{gap}{a character which codes for the gaps (default is "-").}

\item{ambiguous}{a character which codes for the ambiguous state}
}
\value{
The functions return an object of class \code{phyDat}.
}
\description{
The function \code{gap_as_state} changes the contrast of an phyDat object to
treat as its own state.
}
\examples{
data(Laurasiatherian)
tmp <- gap_as_state(Laurasiatherian)
contr <- attr(tmp, "contrast")
rownames(contr) <- attr(tmp, "allLevels")
contr
}
\seealso{
\code{\link{phyDat}}
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
\keyword{cluster}
