% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_phyDat.R
\name{image.phyDat}
\alias{image.phyDat}
\alias{image.ancestral}
\title{Plot of a Sequence Alignment}
\usage{
\method{image}{phyDat}(x, ...)

\method{image}{ancestral}(x, ...)
}
\arguments{
\item{x}{an object containing sequences, an object of class \code{phyDat}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{image.phyDat} returns silently x.
}
\description{
This function plots an image of an alignment of sequences.
}
\details{
A wrapper for using \code{\link[ape]{image.DNAbin}} and
\code{\link[ape]{image.AAbin}}.
Codons triplets are transformed and handled as nucleotide sequences. So far
it is not yet possible to plot data with \code{ type="USER"}.
}
\examples{
data("chloroplast")
image(chloroplast[, 1:50], scheme="Clustal", show.aa = TRUE)
}
\seealso{
\code{\link[ape]{image.DNAbin}}, \code{\link[ape]{image.AAbin}}
}
