% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitch64.R
\name{init_fitch}
\alias{init_fitch}
\title{Initialize data for Fitch parsimony
 
 \code{init_fitch} creates the data structure used from the C++ backend.
 
 @importFrom methods new
 @param obj an alignemt, an object of class \code{phyDat}
 @param parsinfo 
 @return a pointer to the C++ structure}
\usage{
init_fitch(obj, parsinfo = FALSE, order = FALSE, m = 4L, ...)
}
\description{
Initialize data for Fitch parsimony
 
 \code{init_fitch} creates the data structure used from the C++ backend.
 
 @importFrom methods new
 @param obj an alignemt, an object of class \code{phyDat}
 @param parsinfo 
 @return a pointer to the C++ structure
}
