\name{Laurasiatherian}
\alias{Laurasiatherian}
\docType{data}
\title{ Laurasiatherian data (AWCMEE)}
\description{
  Laurasiatherian RNA sequence data 
}
\usage{data(Laurasiatherian)}

\source{
Data have been taken from \url{http://awcmee.massey.ac.nz}
and were converted to R format by \email{K.P.Schliep@massey.ac.nz}.   
}
\examples{
data(Laurasiatherian)
str(Laurasiatherian)
}
\keyword{datasets}
