% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distTree.R
\name{NJ}
\alias{NJ}
\alias{UNJ}
\title{Neighbor-Joining}
\usage{
NJ(x)

UNJ(x)
}
\arguments{
\item{x}{A distance matrix.}
}
\value{
an object of class \code{"phylo"}.
}
\description{
This function performs the neighbor-joining tree estimation of
\bibcitet{Saitou1987, Studier1988}. UNJ is the unweighted version from \bibcitet{Gascuel1997}.
}
\details{
NJ is a wrapper around nj from ape.
}
\examples{

data(Laurasiatherian)
dm <- dist.ml(Laurasiatherian)
tree <- NJ(dm)
plot(tree)

}
\references{
\bibshow{*}
}
\seealso{
\code{\link[ape]{nj}}, \code{\link[ape]{dist.dna}},
\code{\link[phangorn]{dist.hamming}}, \code{\link[phangorn]{upgma}},
\code{\link[ape]{fastme}}
}
\author{
Klaus P. Schliep \email{klaus.schliep@gmail.com}
}
\keyword{cluster}
