% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_ASR.R
\name{parsimony_edgelength}
\alias{parsimony_edgelength}
\title{Assign edge length to tree}
\usage{
parsimony_edgelength(tree, data)
}
\arguments{
\item{tree}{a tree, i.e. an object of class pml}

\item{data}{an object of class phyDat}
}
\value{
a tree with edge length.
}
\description{
\code{parsimony_edgelength} and \code{acctran} assign edge length to a tree where
the edge length is the number of mutations. \code{parsimony_edgelengths}
assigns edge lengths using a joint reconstruction based on the sankoff
algorithm. Ties are broken at random and trees can be multifurating.
\code{acctran} is based on the fitch algorithm and is faster. However trees
need to be bifurcating and ties are split.
}
