% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/anc_heatmap.R, R/assert.R,
%   R/candidate_tree.R, R/hash_phylo.R, R/linkfun.R, R/phangorn-package.R,
%   R/phyDat.R, R/phylo.R, R/plot_networx.R, R/pmlPen.R, R/treeManipulation.R,
%   R/upgma.R
\name{threshStateC}
\alias{threshStateC}
\alias{base_freq}
\alias{reference_position}
\alias{assert_phyDat}
\alias{assert_phylo}
\alias{assert_multiPhylo}
\alias{assert_treeish}
\alias{assert_pml}
\alias{clean_phylo}
\alias{candidate_tree}
\alias{hash}
\alias{hash_topo}
\alias{binomial_mk}
\alias{phangorn-internal}
\alias{coords}
\alias{map_duplicates}
\alias{edQt}
\alias{pml.free}
\alias{pml.init}
\alias{pmlPen}
\alias{relabel}
\alias{edge_length_hclust}
\title{Internal phangorn Functions}
\usage{
threshStateC(x, thresholds)

base_freq(obj)

reference_position(x, pos, ref = 1, gap = "-")

assert_phyDat(x, contains_label = !is.null(label), label = NULL)

assert_phylo(x, has_edge_length = FALSE, is_rooted = FALSE,
  is_ultrametric = FALSE, is_binary = FALSE, unique_tiplabel = TRUE)

assert_multiPhylo(x, has_edge_length = FALSE, is_rooted = FALSE,
  is_ultrametric = FALSE, is_binary = FALSE)

assert_treeish(x, null.ok = FALSE)

assert_pml(x)

clean_phylo(x, ..., unroot = FALSE, collapse.singles = FALSE,
  reorder = FALSE, order = "postorder", multi2di = FALSE,
  di2multi = FALSE, tol = 1e-07, compress = FALSE)

candidate_tree(x, method = c("unrooted", "ultrametric", "tipdated"),
  eps = 1e-08, tip.dates = NULL, ...)

hash(x, ...)

hash_topo(x, rooted = FALSE, ...)

binomial_mk(k = 4)

map_duplicates(x, dist = length(x) < 500, ...)

edQt(Q = c(1, 1, 1, 1, 1, 1), bf = c(0.25, 0.25, 0.25, 0.25))

pml.free()

pml.init(data, k = 1L)

coords(obj, dim = "3D")

pmlPen(object, lambda, ..., control = pml.control())

relabel(x, ref)

edge_length_hclust(x, dm, method = "average")
}
\description{
Internal \pkg{phangorn} functions.
}
\keyword{internal}
