% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phangorn-package.R
\docType{package}
\name{phangorn-package}
\alias{phangorn}
\alias{phangorn-package}
\title{phangorn: Phylogenetic Reconstruction and Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Allows for estimation of phylogenetic trees and networks using Maximum Likelihood, Maximum Parsimony, distance methods and Hadamard conjugation (Schliep 2011). Offers methods for tree comparison, model selection and visualization of phylogenetic networks as described in Schliep et al. (2017).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KlausVigo/phangorn}
  \item \url{https://klausvigo.github.io/phangorn/}
  \item Report bugs at \url{https://github.com/KlausVigo/phangorn/issues}
}

}
\author{
\strong{Maintainer}: Klaus Schliep \email{klaus.schliep@gmail.com} (\href{https://orcid.org/0000-0003-2941-0161}{ORCID})

Authors:
\itemize{
  \item Emmanuel Paradis (\href{https://orcid.org/0000-0003-3092-2199}{ORCID})
  \item Leonardo de Oliveira Martins (\href{https://orcid.org/0000-0001-5247-1320}{ORCID})
  \item Alastair Potts
  \item Iris Bardel-Kahr (\href{https://orcid.org/0000-0002-8950-834X}{ORCID})
}

Other contributors:
\itemize{
  \item Tim W. White [contributor]
  \item Cyrill Stachniss [contributor]
  \item Michelle Kendall \email{m.kendall@imperial.ac.uk} [contributor]
  \item Keren Halabi [contributor]
  \item Richel Bilderbeek [contributor]
  \item Kristin Winchell [contributor]
  \item Liam Revell [contributor]
  \item Mike Gilchrist [contributor]
  \item Jeremy Beaulieu [contributor]
  \item Brian O'Meara [contributor]
  \item Long Qu [contributor]
  \item Joseph Brown (\href{https://orcid.org/0000-0002-3835-8062}{ORCID}) [contributor]
  \item Santiago Claramunt (\href{https://orcid.org/0000-0002-8926-5974}{ORCID}) [contributor]
}

}
\keyword{internal}
