% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnc.R
\name{plotAnc}
\alias{plotAnc}
\title{Plot ancestral character on a tree}
\usage{
plotAnc(tree, data, i = 1, site.pattern = FALSE, col = NULL,
  cex.pie = 0.5, pos = "bottomright", scheme = NULL, ...)
}
\arguments{
\item{tree}{a tree, i.e. an object of class pml}

\item{data}{an object of class \code{phyDat} or \code{ancestral}.}

\item{i}{plots the i-th site of the \code{data}.}

\item{site.pattern}{logical, plot i-th site pattern or i-th site}

\item{col}{a vector containing the colors for all possible states.}

\item{cex.pie}{a numeric defining the size of the pie graphs.}

\item{pos}{a character string defining the position of the legend.}

\item{scheme}{color scheme, for amino acids this can be "Ape_AA",
"Clustal_AA", "Hydrophobicity_AA" and "Zappo_AA" are available.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\description{
\code{plotAnc} plots a phylogeny and adds character to the nodes. Either
takes output from  \code{ancestral.pars} or \code{ancestral.pml} or from an
alignment where there are node labels in the tree match the constructed
sequences in the alignment.
}
\details{
For further details see vignette("Ancestral").
}
\examples{

example(NJ)
# generate node labels to ensure plotting will work
tree <- makeNodeLabel(tree)
anc.p <- ancestral.pars(tree, Laurasiatherian)
# plot the third character
plotAnc(tree, anc.ml, 3)

data(chloroplast)
tree <- pratchet(chloroplast,  maxit=10, trace=0)
tree <- makeNodeLabel(tree)
anc.ch <- ancestral.pars(tree, chloroplast)
image(chloroplast[, 1:25])
plotAnc(tree, anc.ch, 21, scheme="Ape_AA")
plotAnc(tree, anc.ch, 21, scheme="Clustal_AA")
}
\seealso{
\code{\link{ancestral.pml}}, \code{\link[ape]{plot.phylo}}
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
\keyword{plot}
