% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBS.R
\name{plotBS}
\alias{plotBS}
\alias{add_support}
\title{Plotting trees with bootstrap values}
\usage{
plotBS(tree, trees, type = "phylogram", method = "FBP", bs.col = "black",
  bs.adj = NULL, digits = 3, p = 0, frame = "none", tol = 1e-06,
  sep = "/", ...)

add_support(tree, trees, method = "FBP", tol = 1e-08, scale = TRUE,
  frame = "none", digits = 3, sep = "/", ...)
}
\arguments{
\item{tree}{The tree on which edges the bootstrap values are plotted.}

\item{trees}{a list of trees (object of class "multiPhylo").}

\item{type}{the type of tree to plot, one of "phylogram", "cladogram", "fan",
"unrooted", "radial" or "none". If type is "none" the tree is returned with
the bootstrap values assigned to the node labels.}

\item{method}{either "FBP" the classical bootstrap (default), "TBE"
(transfer bootstrap) or "MCC" for assigning clade credibilities. In case of
"MCC" all trees need to be rooted.}

\item{bs.col}{color of bootstrap support labels.}

\item{bs.adj}{one or two numeric values specifying the horizontal and
vertical justification of the bootstrap labels.}

\item{digits}{integer indicating the number of decimal places.}

\item{p}{only plot support values higher than this percentage number
(default is 0).}

\item{frame}{a character string specifying the kind of frame to be printed
around the bootstrap values. This must be one of "none" (the default),
"rect" or "circle".}

\item{tol}{a numeric value giving the tolerance to consider a branch length
significantly greater than zero.}

\item{sep}{seperator between the different methods.}

\item{\dots}{further parameters used by \code{plot.phylo}.}

\item{scale}{return ratio or percentage.}
}
\value{
\code{plotBS} returns silently a tree, i.e. an object of class
\code{phylo} with the bootstrap values as node labels. The argument
\code{trees} is optional and if not supplied the labels supplied
in the \code{node.label} slot will be used.
}
\description{
\code{plotBS} plots a phylogenetic tree with the bootstrap values assigned
to the (internal) edges. It can also used to assign bootstrap values to a
phylogenetic tree. \code{add_support} adds support values to a plot.
}
\details{
The functions can either assign the classical Felsenstein’s bootstrap
proportions (FBP) \bibcitep{Felsenstein1985, Penny1985, Penny1986}, the
transfer bootstrap expectation (TBE) of \bibcitep{Lemoine2018} or "MCC" for
assigning clade credibilities if trees are rooted. Using the option
\code{type=="n"} just assigns the bootstrap values and return the tree
without plotting it.
}
\examples{
fdir <- system.file("extdata/trees", package = "phangorn")
# RAxML best-known tree with bipartition support (from previous analysis)
raxml.tree <- read.tree(file.path(fdir,"RAxML_bipartitions.woodmouse"))
# RAxML bootstrap trees (from previous analysis)
raxml.bootstrap <- read.tree(file.path(fdir,"RAxML_bootstrap.woodmouse"))
par(mfrow=c(1,2))
plotBS(raxml.tree,  raxml.bootstrap, "p")
plotBS(raxml.tree,  raxml.bootstrap, "p", "TBE")
}
\references{
\bibshow{*}
}
\seealso{
\code{\link[ape]{plot.phylo}}, \code{\link{add_ci}},
\code{\link[ape]{nodelabels}},
\code{\link[ape]{prop.clades}}, \code{\link{maxCladeCred}},
\code{\link{transferBootstrap}}, \code{\link[ape]{consensus}},
\code{\link{consensusNet}}
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
