% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_bb.R
\name{pml_bb}
\alias{pml_bb}
\title{Likelihood of a tree.}
\usage{
pml_bb(x, model = NULL, rearrangement = "stochastic",
  method = c("unrooted", "ultrametric", "tip.dated"), bs = 100, ...)
}
\arguments{
\item{x}{An alignment of class (either class \code{phyDat}, \code{DNAbin} or
\code{AAbin}) or an object of class \code{modelTest}.}

\item{model}{A string providing model (e.g. "GTR+G(4)+I") otherwise
\code{\link{modelTest}} is called.}

\item{rearrangement}{Type of tree tree rearrangements to perform, one of
"none", "NNI", "stochastic" or "ratchet"}

\item{method}{Estimate "unrooted", "ultrametric" or "tiplabeled" tree.}

\item{bs}{Number of bootstrap samples.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
\code{pml_bb} returns a list. It contains an object of class
\code{pml}, the best tree found so far and possibly objects of class
\code{modelTest} and the bootstrapped trees.
}
\description{
\code{pml_bb} for pml black box infers a phylogentic tree infers a tree using
maximum likelihood (ML).
}
\details{
It combines the \code{pml}, \code{optim.pml}, \code{modelTest} and
\code{bootstrap.pml}. It first infers the best model, than optimizes this
model and last but not least performs bootstrap analysis.

Currently very experimental and likly to change.
}
\examples{

\dontrun{
data(woodmouse)
tmp <- pml_bb(woodmouse)
}
}
\seealso{
\code{\link{optim.pml}}, \code{\link{bootstrap.pml}},
\code{\link{modelTest}},
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
\keyword{cluster}
