% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terraces.R
\name{terraces}
\alias{terraces}
\title{Discrete Gamma and Beta distribution}
\usage{
terraces(x, bs_trees = NULL, dist_fun = "RF.dist", di2multi = TRUE,
  tol = 2e-08, plot = TRUE, add = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{pml}}

\item{bs_trees}{an object of class \code{multiPhylo}}

\item{dist_fun}{a function to compute distances between trees see e.g.
\code{\link{RF.dist}}}

\item{di2multi}{logical, should trees multichotomies get collapsed.}

\item{tol}{a numeric value giving the tolerance to consider a branch length
significantly greater than zero.}

\item{plot}{loggical if TRUE a 3D scatter is shown.}

\item{add}{whether to add the points to an existing plot.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
\code{terraces} silently returns a matrix.
}
\description{
\code{terraces} visualizes in likelihood surface for the tree space (Sanderson ) Usually trees are from a bootstrap or MCMC sample. There the first two
axis are the principle components of distances between trees and the third
axis is the likelihood value.
}
\examples{
\dontrun{
data(woodmouse)
library(rgl)
fit <- pml_bb(woodmouse, model="JC")
library(rgl)
open3d()
terraces(fit)
}
}
\seealso{
\code{\link{pml_bb}, \link{optim.pml}, \link{pratchet},
\link{RF.dist}, \link[ape]{di2multi}, \link{cmdscale}}.
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
#' @references Sanderson, M.J., McMahon, M.M. and Steel, M. (2011). Terraces in phylogenetic tree space. \emph{Science},
\bold{333}, 448--450.
}
