% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedist.R
\name{treedist}
\alias{treedist}
\alias{sprdist}
\alias{SPR.dist}
\alias{RF.dist}
\alias{wRF.dist}
\alias{KF.dist}
\alias{path.dist}
\title{Distances between trees}
\usage{
treedist(tree1, tree2, check.labels = TRUE)

sprdist(tree1, tree2)

SPR.dist(tree1, tree2 = NULL)

RF.dist(tree1, tree2 = NULL, normalize = FALSE, check.labels = TRUE,
  rooted = FALSE)

wRF.dist(tree1, tree2 = NULL, normalize = FALSE, check.labels = TRUE,
  rooted = FALSE)

KF.dist(tree1, tree2 = NULL, check.labels = TRUE, rooted = FALSE)

path.dist(tree1, tree2 = NULL, check.labels = TRUE, use.weight = FALSE)
}
\arguments{
\item{tree1}{A phylogenetic tree (class \code{phylo}) or vector of trees (an
object of class \code{multiPhylo}). See details}

\item{tree2}{A phylogenetic tree.}

\item{check.labels}{compares labels of the trees.}

\item{normalize}{compute normalized RF-distance, see details.}

\item{rooted}{take bipartitions for rooted trees into account, default is
unrooting the trees.}

\item{use.weight}{use edge.length argument or just count number of edges on
the path (default)}
}
\value{
\code{treedist} returns a vector containing the following tree
distance methods \item{symmetric.difference}{symmetric.difference or
Robinson-Foulds distance}
\item{branch.score.difference}{branch.score.difference}
\item{path.difference}{path.difference}
\item{weighted.path.difference}{weighted.path.difference}
}
\description{
\code{treedist} computes different tree distance methods and \code{RF.dist}
the Robinson-Foulds or symmetric distance. The Robinson-Foulds distance
\bibcitep{Robinson1979, Robinson1981} only depends on the topology of the
trees. If edge weights should be considered  \code{wRF.dist} calculates the
weighted RF distance \bibcitep{Robinson1981} and \code{KF.dist} calculates
the branch score distance \bibcitep{Kuhner1994}.
\code{path.dist} computes the path difference metric as described in
\bibcitet{Steel1993}.
\code{sprdist} computes the approximate SPR distance
\bibcitep{Leo2008, Leo2016}.
}
\details{
The Robinson-Foulds distance between two trees \eqn{T_1} and
\eqn{T_2} with \eqn{n} tips is defined as (following the notation Steel and
Penny 1993):
\deqn{d(T_1, T_2) = i(T_1) + i(T_2) - 2v_s(T_1, T_2)} where \eqn{i(T_1)}
denotes the number of internal edges and \eqn{v_s(T_1, T_2)} denotes the
number of internal splits shared by the two trees. The normalized
Robinson-Foulds distance is derived by dividing \eqn{d(T_1, T_2)} by the
maximal possible distance \eqn{i(T_1) + i(T_2)}. If both trees are unrooted
and binary this value is \eqn{2n-6}.

Functions like \code{RF.dist} returns the Robinson-Foulds distance between
either 2 trees or computes a matrix of all pairwise distances if a
\code{multiPhylo} object is given.

For large number of trees the distance functions can use a lot of memory!
}
\examples{

tree1 <- rtree(100, rooted=FALSE)
tree2 <- rSPR(tree1, 3)
RF.dist(tree1, tree2)
treedist(tree1, tree2)
sprdist(tree1, tree2)
trees <- rSPR(tree1, 1:5)
SPR.dist(tree1, trees)

}
\references{
\bibshow{*}
}
\seealso{
\code{\link[ape]{dist.topo}}, \code{\link{nni}},
\code{\link{superTree}}, \code{\link{mast}}
}
\author{
Klaus P. Schliep \email{klaus.schliep@gmail.com},
Leonardo de Oliveira Martins
}
\keyword{classif}
