% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestral.R
\name{write.ancestral}
\alias{write.ancestral}
\alias{ancestral}
\alias{print.ancestral}
\title{Export and convenience functions for  ancestral reconstructions}
\usage{
write.ancestral(x, file = "ancestral")

ancestral(tree, align, prob)

\method{print}{ancestral}(x, ...)
}
\arguments{
\item{x}{an object of class ancestral.}

\item{file}{a file name. File endings are added.}

\item{tree}{an object of class phylo.}

\item{align}{an object of class phyDat.}

\item{prob}{an data.frame containing a matrix of posterior probabilities for
each state and site.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
\code{write.ancestral} allows to export ancestral reconstructions. It writes
out the tree, a tab delimited text file with the probabilities and the
alignment.  \code{read.ancestral} tries to read in these files.
\code{ancestral} generates an object of class ancestral.
}
\details{
This allows also to read in reconstruction  made by iqtree to use the
plotting capabilities of R.
}
