% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_generics.R
\name{write.pml}
\alias{write.pml}
\title{Export pml objects}
\usage{
write.pml(x, file = "pml", save_rds = FALSE, digits = 10, ...)
}
\arguments{
\item{x}{an object of class ancestral.}

\item{file}{a file name. File endings are added.}

\item{save_rds}{logical, if TRUE saves the pml object as a rds file,
otherwise the alignment is saved as a fasta file.}

\item{digits}{default is 10, i.e. edge length for the bootstrap trees are
exported. For digits larger smaller than zero no edge length are exported.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\code{write.pml}  returns the input x invisibly.
}
\description{
\code{write.pml} writes out the ML tree and the model parameters.
}
\details{
\code{write.pml} creates several files. It exports the alignment as fasta
file. It writes out the ML tree in a newick file and the estimates parameters
in a txt file.  It should be possible to (re-)create the pml object up to
numerical inaccuracies.
If bootstrap trees exist these are additionally exported in a nexus file.
}
\examples{
data(woodmouse)
fit <- pml_bb(woodmouse, "JC", rearrangement = "none")
write.pml(fit, "woodmouse")
unlink(c("woodmouse.txt", "woodmouse_tree.nwk", "woodmouse_align.fasta"))
}
\seealso{
\code{\link{ancestral.pml}}, \code{\link{plotAnc}}
}
