% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phangorn-package.R
\docType{data}
\name{yeast}
\alias{yeast}
\title{Yeast alignment (Rokas et al.)}
\description{
Alignment of 106 genes of 8 different species of yeast.
}
\examples{

data(yeast)
str(yeast)

}
\references{
Rokas, A., Williams, B. L., King, N., and Carroll, S. B. (2003)
Genome-scale approaches to resolving incongruence in molecular phylogenies.
\emph{Nature}, \bold{425}(6960): 798--804
}
\keyword{datasets}
