% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_CI.R
\name{add_edge_length}
\alias{add_edge_length}
\title{Assign and compute edge lengths from a sample of trees}
\usage{
add_edge_length(tree, trees, fun = function(x) median(na.omit(x)),
  rooted = TRUE)
}
\arguments{
\item{tree}{tree where edge lengths are assigned to.}

\item{trees}{an object of class multiPhylo, where the average for the edges
is computed from.}

\item{fun}{a function to compute the average (default is median).}

\item{rooted}{the height of the boxes.}
}
\description{
This command can infer some average edge lengths and assign
them from a (bootstrap/MCMC) sample.
}
\examples{
\dontrun{
data("Laurasiatherian")
dm <- dist.hamming(Laurasiatherian)
set.seed(123)
trees <- bootstrap.phyDat(Laurasiatherian,
                          FUN=function(x)upgma(dist.hamming(x)), bs=100)
                          tree <- plotBS(tree, trees, "phylogram")
tree_ultra <- allCompat(trees, rooted=TRUE) |> add_edge_length(trees)
tree_unrooted <- allCompat(trees, rooted=FALSE) |>
                 add_edge_length(trees, rooted=FALSE)
plot(tree_ultra)
plot(tree_unrooted)
}
}
\seealso{
\code{\link{node.depth.edgelength}}, \code{\link{consensus}},
\code{\link{maxCladeCred}}
}
\author{
Klaus Schliep
}
\keyword{aplot}
