% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anc_heatmap.R
\name{anc_heatmap}
\alias{anc_heatmap}
\title{anc_heatmap}
\usage{
anc_heatmap(x, y = NULL, use.edge.length = FALSE, align_label = TRUE,
  clade = NULL, select = NULL, ...)
}
\arguments{
\item{x}{an object of class ancestral or phyDat.}

\item{y}{an object  of class phyDat.}

\item{use.edge.length}{a logical indicating whether to use the edge lengths
of the phylogeny to draw the branches (the default) or not (if FALSE). This
option has no effect if the object of class "phylo" has no ‘edge.length’
element.}

\item{align_label}{a logical value or an integer. If TRUE, the tips are
aligned and dotted lines are drawn between the nodes of the tree and the
labels.}

\item{clade}{a node number or label to extract the clade from the tree.}

\item{select}{a subset of characters, columns in the alignment.}
}
\description{
Plots a heat map with ancestral states.
}
\details{
\code{anc_heatmap} plot the joint distribution or the most likely state.
}
\examples{
data(woodmouse)
wm <- as.phyDat(woodmouse)
tree <- pratchet(wm)
tree <- makeNodeLabel(tree)
anc_mp <- anc_pars(tree, wm)
anc_heatmap(anc_mp)
}
\seealso{
\code{\link{anc_pml}}, \code{\link{plotAnc}}
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
