% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networx.R
\name{as.splits}
\alias{addTrivialSplits}
\alias{allCircularSplits}
\alias{allSplits}
\alias{as.Matrix}
\alias{as.Matrix.splits}
\alias{as.bitsplits.splits}
\alias{as.matrix.splits}
\alias{as.phylo.splits}
\alias{as.prop.part.splits}
\alias{as.splits}
\alias{as.splits.multiPhylo}
\alias{as.splits.networx}
\alias{as.splits.phylo}
\alias{c.splits}
\alias{compatible}
\alias{countCycles}
\alias{distinct.splits}
\alias{matchSplits}
\alias{print.splits}
\alias{removeTrivialSplits}
\alias{unique.splits}
\title{Splits representation of graphs and trees.}
\usage{
as.splits(x, ...)

\method{print}{splits}(x, maxp = getOption("max.print"), zero.print = ".",
  one.print = "|", ...)

\method{as.splits}{phylo}(x, ...)

\method{as.splits}{multiPhylo}(x, ...)

\method{as.prop.part}{splits}(x, ...)

compatible(obj)

allSplits(k, labels = NULL)

allCircularSplits(k, labels = NULL)
}
\arguments{
\item{x}{An object of class phylo or multiPhylo.}

\item{maxp}{integer, default from \code{options(max.print)}, influences how
many entries of large matrices are printed at all.}

\item{zero.print}{character which should be printed for zeros.}

\item{one.print}{character which should be printed for ones.}

\item{obj}{an object of class splits.}

\item{k}{number of taxa.}

\item{labels}{names of taxa.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
\code{as.splits} returns an object of class splits, which is mainly
a list of splits and some attributes. Often a \code{splits} object will
contain attributes \code{confidences} for bootstrap or Bayesian support
values and \code{weight} storing edge weights.  
\code{compatible} return a lower triangular matrix where an 1 indicates that
two splits are incompatible.
}
\description{
\code{as.splits} produces a list of splits or bipartitions.
}
\note{
The internal representation is likely to change.
}
\examples{

(sp <- as.splits(rtree(5)))
write.nexus.splits(sp)
spl <- allCircularSplits(5)
plot(as.networx(spl), "2D")

}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
\seealso{
\code{\link{prop.part}}, \code{\link{lento}}, \code{\link{as.networx}}, 
\code{\link{distanceHadamard}}, \code{\link{read.nexus.splits}}
}
\keyword{cluster}

