% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phangorn-package.R
\docType{data}
\name{chloroplast}
\alias{chloroplast}
\title{Chloroplast alignment}
\description{
Amino acid alignment of 15 genes of 19 different chloroplast.
}
\examples{

data(chloroplast)
chloroplast

}
\keyword{datasets}
