% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonTest.R
\name{codonTest}
\alias{codonTest}
\title{codonTest}
\usage{
codonTest(tree, object, model = c("M0", "M1a", "M2a"),
  frequencies = "F3x4", opt_freq = FALSE, codonstart = 1, ...)
}
\arguments{
\item{tree}{a phylogenetic tree.}

\item{object}{an object of class phyDat.}

\item{model}{a vector containing the substitution models to compare with
each other or "all" to test all available models.}

\item{frequencies}{a character string or vector defining how to compute
the codon frequencies}

\item{opt_freq}{optimize frequencies (so far ignored)}

\item{codonstart}{an integer giving where to start the translation. This
should be 1, 2, or 3, but larger values are accepted and have for effect to
start the translation further within the sequence.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A data.frame containing the log-likelihood, number of estimated
parameters, AIC, AICc and BIC all tested models.  The data.frame has an
attributes "env" which is an environment which contains all the trees, the
data and the calls to allow get the estimated models, e.g. as a starting
point for further analysis (see example).
}
\description{
Models for detecting positive selection
}
\details{
Currently \code{codonTest} estimates all the specified models for a given
tree and data.
There are several options for deriving the codon frequencies.
Frequencies can be "equal" (1/61), derived from nucleotide frequencies "F1x4"
and "F3x4" or "empirical" codon frequencies. The frequencies taken using
the empirical frequencies or estimated via maximum likelihood.

So far the M0 model (Goldman and Yang 2002), the M1, M1a and M2a are
implemented. The M0 model is always computed the other a re optional.
M1a can be used as the null model to test positive selection
}
\references{
Ziheng Yang (2014). \emph{Molecular Evolution: A Statistical
Approach}. Oxford University Press, Oxford
}
\seealso{
\code{\link{pml}}, \code{\link{pmlMix}}, \code{\link{modelTest}},
\code{\link[stats]{AIC}}
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
\keyword{cluster}
