% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedist.R
\name{cophenetic.networx}
\alias{cophenetic.networx}
\alias{cophenetic.splits}
\title{Pairwise Distances from a Phylogenetic Network}
\usage{
\method{cophenetic}{networx}(x)
}
\arguments{
\item{x}{an object of class \code{networx}.}
}
\value{
an object of class \code{dist}, names are set according to the tip
labels (as given by the element \code{tip.label} of the argument \code{x}).
}
\description{
\code{cophenetic.networx} computes the pairwise distances between the pairs
of tips from a phylogenetic network using its branch lengths.
}
\examples{
example(neighborNet)
cophenetic(nnet)
}
\seealso{
\code{\link[stats]{cophenetic}} for the generic function,
\code{neighborNet} to construct a network from a distance matrix
}
\author{
Klaus Schliep
}
\keyword{manip}
