% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadamard.R
\name{distanceHadamard}
\alias{distanceHadamard}
\title{Distance Hadamard}
\usage{
distanceHadamard(dm, eps = 0.001)
}
\arguments{
\item{dm}{A distance matrix.}

\item{eps}{Threshold value for splits.}
}
\value{
\code{distanceHadamard} returns a matrix. The first column contains
the distance spectra, the second one the edge-spectra. If eps is positive an
object of with all splits greater eps is returned.
}
\description{
Distance Hadamard produces spectra of splits from a distance matrix.
}
\examples{

data(yeast)
dm <- dist.hamming(yeast)
dm <- as.matrix(dm)
fit <- distanceHadamard(dm)
lento(fit)
plot(as.networx(fit), "2D")

}
\references{
Hendy, M. D. and Penny, D. (1993). Spectral Analysis of
Phylogenetic Data. \emph{Journal of Classification}, \bold{10}, 5-24.
}
\seealso{
\code{\link{hadamard}}, \code{\link{lento}},
\code{\link{plot.networx}}, \code{\link{neighborNet}}
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}, Tim White
}
\keyword{cluster}
