% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_phyDat.R
\name{image.phyDat}
\alias{image.phyDat}
\title{Plot of a Sequence Alignment}
\usage{
\method{image}{phyDat}(x, ...)
}
\arguments{
\item{x}{an object containing sequences, an object of class \code{phyDat}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function plots an image of an alignment of sequences.
}
\details{
A wrapper for using \code{\link{image.DNAbin}} and \code{\link{image.AAbin}}.
}
\seealso{
\code{\link{image.DNAbin}}, \code{\link{image.AAbin}}
}
