% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_phyDat.R
\name{image.phyDat}
\alias{image.phyDat}
\title{Plot of a Sequence Alignment}
\usage{
\method{image}{phyDat}(x, ...)
}
\arguments{
\item{x}{an object containing sequences, an object of class \code{phyDat}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Nothing. The function is called for plotting.
}
\description{
This function plots an image of an alignment of sequences.
}
\details{
A wrapper for using \code{\link[ape]{image.DNAbin}} and \code{\link[ape]{image.AAbin}}.
Codons triplets are handled as nucleotide sequences.
}
\examples{
data("chloroplast")
image(chloroplast[, 1:50], scheme="Clustal", show.aa = TRUE)
}
\seealso{
\code{\link[ape]{image.DNAbin}}, \code{\link[ape]{image.AAbin}}
}
