\name{Laurasiatherian}
\alias{Laurasiatherian}
\docType{data}
\title{ Laurasiatherian data (AWCMEE)}
\description{
  Laurasiatherian RNA sequence data 
}
\usage{data(Laurasiatherian)}

\source{
Data have been taken from \url{http://www.allanwilsoncentre.ac.nz/}
and were converted to R format by \email{klaus.schliep@gmail.com}.   
}
\examples{
data(Laurasiatherian)
str(Laurasiatherian)
}
\keyword{datasets}
