% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phangorn-package.R
\docType{data}
\name{Laurasiatherian}
\alias{Laurasiatherian}
\title{Laurasiatherian data (AWCMEE)}
\source{
Data have been taken from the former repository of the Allan Wilson
Centre and were converted to R format by \email{klaus.schliep@gmail.com}.
}
\description{
Laurasiatherian RNA sequence data
}
\examples{

data(Laurasiatherian)
str(Laurasiatherian)

}
\keyword{datasets}
