% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTest.R
\name{modelTest}
\alias{modelTest}
\alias{AICc}
\alias{write.modelTest}
\title{ModelTest}
\usage{
modelTest(object, tree = NULL, model = NULL, G = TRUE, I = TRUE,
  FREQ = FALSE, k = 4, control = pml.control(), RHAS = c("gamma",
  "gamma_weighted", "free_rate"), ..., mt_control = mt.control(crit = "BIC",
  n_model = 5, n_rhas = 100))

write.modelTest(x, file = "modelTest", digits = 10)
}
\arguments{
\item{object}{an object of class phyDat or pml.}

\item{tree}{a phylogenetic tree.}

\item{model}{a vector containing the substitution models to compare with
each other or "all" to test all available models.}

\item{G}{logical, TRUE (default) if (discrete) Gamma model should be tested.}

\item{I}{logical, TRUE (default) if invariant sites should be tested.}

\item{FREQ}{logical, FALSE (default) if TRUE amino acid frequencies will be
estimated.}

\item{k}{number of rate classes. Can be a list with a vector for each RHAS
term.}

\item{control}{A list of parameters for controlling the fitting process.}

\item{RHAS}{a character vector specifying the rate heterogeneity among sites
models. Option are "gamma" for discrete gamma model with equal weights,
"gamma_weighted" for discrete gamma model with estimated weights,
"free_rate" and "gamma_quadrature".}

\item{...}{Further arguments passed to or from other methods.}

\item{mt_control}{a list with some options.}

\item{x}{an object of class modelTest.}

\item{file}{a file name. File endings are added.}

\item{digits}{default is 10, i.e. edge length for the bootstrap trees are
exported. For digits larger smaller than zero no edge length are exported.}
}
\value{
A data.frame containing the log-likelihood, number of estimated
parameters, AIC, AICc and BIC all tested models.  The data.frame has an
attributes "env" which is an environment which contains all the trees, the
data and the calls to allow get the estimated models, e.g. as a starting
point for further analysis (see example).
}
\description{
Comparison of different nucleotide or amino acid substitution models
}
\details{
\code{modelTest} estimates all the specified models for a given tree and
data similar to \bibcitep{Posada1998, Posada2008}.

model="3Di" tests the three models for structural alignments mention in
\bibcitet{Garg2025}.
}
\examples{

\dontrun{
data(Laurasiatherian)
mT <- modelTest(Laurasiatherian, model = c("JC", "K80", "HKY", "GTR"))

# Some exploratory data analysis
plot(mT$TL, mT$logLik, xlim=c(3,6.5),
   xlab="total tree length", ylab="log likelihood")
text(mT$TL, mT$logLik, labels=mT$Model, pos=4)

fit_GTR_G <- as.pml(mt, "GTR+G(4)")
fit_GTR_GW <- as.pml(mt, "GTR+GW(4)")
fit_GTR_R <- as.pml(mt, "GTR+R(4)")
plotRates(fit_GTR_G)
plotRates(fit_GTR_GW, append=TRUE, col="red")
plotRates(fit_GTR_R, append=TRUE, col="green")

# extract best model
(best_model <- as.pml(mT))


data(chloroplast)
(mTAA <- modelTest(chloroplast, model=c("JTT", "WAG", "LG")))

# test all available amino acid models
plan(multisession, workers = 2)
(mTAA_all <- modelTest(chloroplast, model="all"))
plan(sequential)
}

}
\references{
\bibshow{*}

Burnham, K. P. and Anderson, D. R (2002) \emph{Model selection
and multimodel inference: a practical information-theoretic approach}. 2nd
ed. Springer, New York

Darriba D., Taboada G.L., Doallo R and Posada D. (2011) ProtTest 3: fast
selection of best-fit models of protein evolution. . \emph{Bioinformatics}
\bold{27}: 1164-1165

Puente-Lelievre, Caroline, et al. (2023) "Tertiary-interaction characters
enable fast, model-based structural phylogenetics beyond the twilight zone."
\emph{bioRxiv}: 2023-12
}
\seealso{
\code{\link{pml}}, \code{\link{anova}}, \code{\link[stats]{AIC}},
\code{\link{codonTest}}, \code{\link[future]{plan}}
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
\keyword{cluster}
