\name{phangorn-internal}
\alias{phangorn-internal}
\alias{fast.table}
\alias{optEdge}
\alias{fs}
\alias{optimQ}
\alias{optimInv}
\alias{optimGamma}
\alias{optimEdge}
\alias{optimW}
\alias{phyloNNI}
\alias{dist.logDet}
\alias{dist.hamming}
\alias{fn.quartet}
\alias{fn.triplet}
\alias{optim.quartet}
\alias{optim.triplet}
\alias{optimBf}
\alias{pml.nni}
\alias{pml2}
\alias{lli}
\alias{ll}
\alias{rnodes}
\alias{edQ}
\alias{edQt}
\alias{getP}
\alias{score}
\alias{generateSP}
\alias{write.phylip}
\alias{pml3}
\alias{ll2}
\alias{nnimove}
\alias{pml4}
\alias{baseFreq}
\alias{fit.sankoff}
\alias{getdP}
\alias{getdP2}
\alias{getRows}
\alias{pml5}
\alias{pnodes}
\alias{prepareSankoff}
\alias{sankoff.nni}
\alias{sankoff.quartet}
\alias{sankoffNNI}
\alias{write.nxs.splits}
\alias{bipartition}
\alias{prepareDataSankoff}
\alias{incomplete.gamma}
\alias{discrete.gamma}

\title{Internal distpars functions}
\description{
  Internal distpars functions.
}
\details{
  These are not to be called by the user or are not documented yet.
}
\keyword{internal}

