% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/distTree.R, R/networx.R,
%   R/phangorn-package.R, R/phyDat.R
\name{threshStateC}
\alias{threshStateC}
\alias{PNJ}
\alias{coords}
\alias{phangorn-internal}
\alias{map_duplicates}
\title{Internal phangorn Functions}
\usage{
threshStateC(x, thresholds)

PNJ(data)

coords(obj, dim = "3D")

map_duplicates(x, dist = TRUE, ...)
}
\description{
Internal \pkg{phangorn} functions.
}
\keyword{internal}
