% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/candidate_tree.R,
%   R/hash_phylo.R, R/phangorn-package.R, R/phyDat.R, R/phylo.R,
%   R/plot_networx.R, R/pmlPen.R, R/treeManipulation.R
\name{threshStateC}
\alias{threshStateC}
\alias{candidate_tree}
\alias{hash}
\alias{phangorn-internal}
\alias{coords}
\alias{map_duplicates}
\alias{edQt}
\alias{pml.free}
\alias{pml.init}
\alias{pmlPen}
\alias{relabel}
\title{Internal phangorn Functions}
\usage{
threshStateC(x, thresholds)

candidate_tree(x, method = c("unrooted", "ultrametric", "tipdated"),
  eps = 1e-08, tip.dates = NULL, ...)

hash(x, ...)

map_duplicates(x, dist = length(x) < 500, ...)

edQt(Q = c(1, 1, 1, 1, 1, 1), bf = c(0.25, 0.25, 0.25, 0.25))

pml.free()

pml.init(data, k = 1L)

coords(obj, dim = "3D")

pmlPen(object, lambda, ...)

relabel(x, ref)
}
\description{
Internal \pkg{phangorn} functions.
}
\keyword{internal}
