% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/phangorn-package.R
\name{threshStateC}
\alias{threshStateC}
\alias{phangorn-internal}
\alias{coords}
\alias{map_duplicates}
\title{Internal phangorn Functions}
\usage{
threshStateC(x, thresholds)
}
\description{
Internal \pkg{phangorn} functions.
}
\keyword{internal}
