% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/assert.R, R/candidate_tree.R,
%   R/hash_phylo.R, R/phangorn-package.R, R/phyDat.R, R/phylo.R,
%   R/plot_networx.R, R/pmlPen.R, R/treeManipulation.R, R/upgma.R
\name{threshStateC}
\alias{threshStateC}
\alias{assert_phylo}
\alias{assert_treeish}
\alias{assert_multiPhylo}
\alias{assert_phyDat}
\alias{assert_pml}
\alias{assert_dist}
\alias{clean_phylo}
\alias{clean_multiPhylo}
\alias{candidate_tree}
\alias{hash}
\alias{phangorn-internal}
\alias{coords}
\alias{map_duplicates}
\alias{edQt}
\alias{pml.free}
\alias{pml.init}
\alias{pmlPen}
\alias{relabel}
\alias{edge_length_hclust}
\title{Internal phangorn Functions}
\usage{
threshStateC(x, thresholds)

assert_phylo(x, has_edge_length = FALSE, is_rooted = FALSE,
  is_ultrametric = FALSE)

assert_treeish(x, null.ok = FALSE)

assert_multiPhylo(x, is_ultrametric = FALSE, is_rooted = FALSE,
  same_labels = FALSE, has_edge_length = FALSE)

assert_phyDat(x, contains_label = !is.null(label), label = NULL)

assert_pml(x)

assert_dist(x, finite = FALSE, missing = FALSE)

clean_phylo(x, unroot = FALSE, multi2di = FALSE,
  collapse.singles = FALSE, reorder = FALSE)

clean_multiPhylo(x, unroot = FALSE, multi2di = FALSE,
  collapse.singles = FALSE, reorder = FALSE, compress = FALSE)

candidate_tree(x, method = c("unrooted", "ultrametric", "tipdated"),
  eps = 1e-08, tip.dates = NULL, ...)

hash(x, ...)

map_duplicates(x, dist = length(x) < 500, ...)

edQt(Q = c(1, 1, 1, 1, 1, 1), bf = c(0.25, 0.25, 0.25, 0.25))

pml.free()

pml.init(data, k = 1L)

coords(obj, dim = "3D")

pmlPen(object, lambda, ...)

relabel(x, ref)

edge_length_hclust(x, dm, method = "average")
}
\description{
Internal \pkg{phangorn} functions.
}
\keyword{internal}
