% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/candidate_tree.R,
%   R/hash_phylo.R, R/networx.R, R/phangorn-package.R, R/phyDat.R, R/pmlPen.R
\name{threshStateC}
\alias{threshStateC}
\alias{candidate_tree}
\alias{hash_phylo}
\alias{coords}
\alias{phangorn-internal}
\alias{map_duplicates}
\alias{pmlPen}
\title{Internal phangorn Functions}
\usage{
threshStateC(x, thresholds)

candidate_tree(x, method = c("unrooted", "ultrametric", "tipdated"),
  eps = 1e-08, tip.dates = NULL, ...)

hash_phylo(x, ...)

coords(obj, dim = "3D")

map_duplicates(x, dist = length(x) < 500, ...)

pmlPen(object, lambda, ...)
}
\description{
Internal \pkg{phangorn} functions.
}
\keyword{internal}
