% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/networx.R,
%   R/phangorn-package.R, R/phyDat.R, R/pmlPen.R
\name{threshStateC}
\alias{threshStateC}
\alias{coords}
\alias{phangorn-internal}
\alias{map_duplicates}
\alias{pmlPen}
\title{Internal phangorn Functions}
\usage{
threshStateC(x, thresholds)

coords(obj, dim = "3D")

map_duplicates(x, dist = TRUE, ...)

pmlPen(object, lambda, ...)
}
\description{
Internal \pkg{phangorn} functions.
}
\keyword{internal}
