% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnc.R
\name{plot.ancestral}
\alias{plot.ancestral}
\alias{plotAnc}
\alias{plotSeqLogo}
\alias{image.ancestral}
\title{Plot ancestral character on a tree}
\usage{
\method{plot}{ancestral}(x, which = c(1, 2, 3), site = 1, node = NULL,
  col = NULL, cex.pie = 0.5, pos = "bottomright", scheme = NULL,
  start = 1, end = 10, ...)

plotAnc(x, i = 1, col = NULL, cex.pie = 0.5, pos = "bottomright",
  scheme = NULL, ...)

plotSeqLogo(x, node, start = 1, end = 10, scheme = "Ape_NT", ...)

\method{image}{ancestral}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ancestral}.}

\item{which}{a subset of the numbers 1:3, by default 1:3, referring to
\enumerate{
\item "tree with pie charts" plot
\item "seqlogo" plot
\item "image" plot
}}

\item{node}{to plot for which the propabilities should be plotted.}

\item{col}{a vector containing the colors for all possible states.}

\item{cex.pie}{a numeric defining the size of the pie graphs.}

\item{pos}{a character string defining the position of the legend.}

\item{scheme}{a predefined color scheme. For amino acid options are "Ape_AA",
"Zappo_AA", "Clustal", "Polarity" and "Transmembrane_tendency", for
nucleotides "Ape_NT" and"RY_NT". Names can be abbreviated.}

\item{start}{start position to plot.}

\item{end}{end position to plot.}

\item{\dots}{Further arguments passed to or from other methods.}

\item{i, site}{plots the i-th site.}
}
\description{
\code{plotAnc} plots a phylogeny and adds character to the nodes. Either
takes output from  \code{ancestral.pars} or \code{ancestral.pml} or from an
alignment where there are node labels in the tree match the constructed
sequences in the alignment.
}
\details{
For further details see vignette("Ancestral").
}
\examples{

example(NJ)
# generate node labels to ensure plotting will work
tree <- makeNodeLabel(tree)
anc.p <- ancestral.pars(tree, Laurasiatherian)
# plot the third character
## plotAnc(tree, anc.p, 3)
plotAnc(anc.p, 3)
plotSeqLogo(anc.p, node="Node10", 1, 25)

data(chloroplast)
tree <- pratchet(chloroplast,  maxit=10, trace=0)
tree <- makeNodeLabel(tree)
anc.ch <- ancestral.pars(tree, chloroplast)
image(chloroplast[, 1:25])
plotAnc(anc.ch, 21, scheme="Ape_AA")
plotAnc(anc.ch, 21, scheme="Clustal")
plotSeqLogo(anc.ch, node="Node1", 1, 25, scheme="Clustal")
}
\seealso{
\code{\link{ancestral.pml}}, \code{\link[ape]{plot.phylo}},
\code{\link[ape]{image.DNAbin}}, \code{\link[ape]{image.AAbin}}
\code{\link[ggseqlogo]{ggseqlogo}}
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
\keyword{plot}
