% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pml.R
\name{plot.pml}
\alias{plot.pml}
\title{Plot phylogeny of a pml object}
\usage{
\method{plot}{pml}(x, type = "phylogram", ...)
}
\arguments{
\item{x}{an object of class \code{pml} or \code{phyDat}.}

\item{type}{a character string specifying the type of phylogeny to be drawn;
it must be one of "phylogram" (the default), "cladogram", "fan", "unrooted",
"radial", "tidy", or any unambiguous abbreviation of these.}

\item{\dots}{further parameters to be passed to \code{plot.phylo}.}
}
\description{
\code{plot.pml} is a warapper around \code{plot.phylo} with different default
values for unrooted, ultrametric and tip dated phylogenies.
}
\examples{
fdir <- system.file("extdata/trees", package = "phangorn")
tmp <- read.csv(file.path(fdir,"H3N2_NA_20.csv"))
H3N2 <- read.phyDat(file.path(fdir,"H3N2_NA_20.fasta"), format="fasta")
dates <- setNames(tmp$numdate_given, tmp$name)

fit_td <- pml_bb(H3N2, model="JC", method="tipdated", tip.dates=dates,
                 rearrangement="none", control = pml.control(trace = 0))
plot(fit_td, show.tip.label = FALSE)
# Same as:
# root_time <- max(dates) - max(node.depth.edgelength(fit_td$tree))
# plot(fit_td$tree, show.tip.label = FALSE)
# axisPhylo(root.time = root_time, backward = FALSE)

fit_unrooted <- pml_bb(H3N2, model="JC", rearrangement="none",
                       control = pml.control(trace = 0))
plot(fit_unrooted, cex=.5)

}
\seealso{
\code{\link{plot.phylo}}, \code{\link{axisPhylo}},
\code{\link{add.scale.bar}}
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
\keyword{plot}
