% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terraces.R
\name{terraces}
\alias{terraces}
\alias{terraces.pml}
\alias{terraces.phyDat}
\title{Explore likelihood parsimony surface}
\usage{
terraces(x, ...)

\method{terraces}{pml}(x, trees = x$bs, dist_fun = "RF.dist",
  di2multi = FALSE, tol = 2e-08, plot = TRUE, ...)

\method{terraces}{phyDat}(x, trees, dist_fun = "RF.dist", di2multi = TRUE,
  tol = 2e-08, plot = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{pml}}

\item{\dots}{Further arguments passed to or from other methods.}

\item{trees}{an object of class \code{multiPhylo}}

\item{dist_fun}{a function to compute distances between trees see e.g.
\code{\link{RF.dist}}}

\item{di2multi}{logical, should trees multichotomies get collapsed. Useful
for Robinson-Foulds distance. If edge length are used to compute the
distance, e.g. Kuhner-Felsenstein distance, this is not needed.}

\item{tol}{a numeric value giving the tolerance to consider a branch length
significantly greater than zero.}

\item{plot}{loggical if TRUE a 3D scatter is shown.}
}
\value{
\code{terraces} silently returns a matrix.
}
\description{
\code{terraces} visualizes in likelihood surface for the tree space
(Sanderson et al. 2011). Usually trees are from a bootstrap or MCMC sample.
There the first two axis are the principle components of distances between
trees and the third axis is the likelihood value or parsimony score.
}
\examples{
data(woodmouse)
pratchet(woodmouse)
trs <- pratchet(woodmouse, all=TRUE)
start_trs <- get("start_trees", envir = attr(trs, "env"))
terraces(woodmouse, c(trs, start_trs))

\dontrun{
fit <- pml_bb(woodmouse, model="JC")
terraces(fit, dist_fun="KF.dist")
terraces(fit, pkg="scatterplot3d")
terraces(fit, pkg="plot3D")
terraces(fit, pkg="rgl")
}
}
\references{
Sanderson, M.J., McMahon, M.M. and Steel, M. (2011). Terraces in
phylogenetic tree space. \emph{Science},
\bold{333}, 448--450.
}
\seealso{
\code{\link{pml_bb}, \link{optim.pml}, \link{pratchet},
\link{RF.dist}, \link[ape]{di2multi}, \link{cmdscale}}.
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
