% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pml_generics.R
\name{write.pml}
\alias{write.pml}
\title{Export pml objects}
\usage{
write.pml(x, file = "pml", save_rds = TRUE, ...)
}
\arguments{
\item{x}{an object of class ancestral.}

\item{file}{a file name. File endings are added.}

\item{save_rds}{logical, if TRUE saves the pml object as a rds file,
otherwise the alignment is saved as a fasta file.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\code{write.pml}  returns the input x invisibly.
}
\description{
\code{write.pml} writes out the ML tree and the model parameters.
}
\examples{
data(woodmouse)
fit <- pml_bb(woodmouse, "JC", rearrangement = "none")
write.pml(fit, "woodmouse")
unlink(c("woodmouse_pml.txt", "woodmouse_tree.nwk", "woodmouse.rds"))
}
\seealso{
\code{\link{ancestral.pml}}, \code{\link{plotAnc}}
}
