\name{yeast}
\alias{yeast}
\docType{data}
\title{ Yeast alignment (Rokas et al.) }
\description{
Alignment of 106 genes of 8 different species of yeast.}
\usage{data(yeast)}

\references{
Rokas, A., Williams, B. L., King, N., and Carroll, S. B. (2003)
Genome-scale approaches to resolving incongruence in molecular phylogenies.
\emph{Nature}, \bold{425}(6960): 798--804
}    

\examples{
data(yeast)
str(yeast)
}
\keyword{datasets}
